svn merge -r5669:9767 file:///noble/svn/crux/branches/crux-pwiz

+++ src/c/Scorer.cpp
@@ -1631,7 +1631,7 @@ double score_logp_bonf_weibull(
     return 0.0;
   }
   else {
-    double p_value = exp(-pow(temp/eta, beta));
+    double p_value = exp(-pow((double) temp/eta, (double) beta));
     carp(CARP_DETAILED_DEBUG, "Stat: pvalue before = %.15f", p_value);
 
     // The Bonferroni correction
